moved to programatic approach
