--- remotewrite/sender/go.mod
@@ -1,27 +1,0 @@
-module github.com/prometheus/compliance/remotewrite/sender
-
-go 1.25.0
-
-require (
-	github.com/golang/snappy v1.0.0
-	github.com/oklog/run v1.2.0
-	github.com/prometheus/client_golang/exp v0.0.0-20250914183048-a974e0d45e0a
-	github.com/prometheus/prometheus v0.307.4-0.20251119130332-1174b0ce4f1f
-	github.com/stretchr/testify v1.11.1
-	gopkg.in/yaml.v3 v3.0.1
-)
-
-require (
-	github.com/cespare/xxhash/v2 v2.3.0 // indirect
-	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
-	github.com/gogo/protobuf v1.3.2 // indirect
-	github.com/grafana/regexp v0.0.0-20250905093917-f7b3be9d1853 // indirect
-	github.com/klauspost/compress v1.18.1 // indirect
-	github.com/kr/pretty v0.3.1 // indirect
-	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
-	github.com/prometheus/client_model v0.6.2 // indirect
-	github.com/prometheus/common v0.67.2 // indirect
-	go.yaml.in/yaml/v2 v2.4.3 // indirect
-	golang.org/x/text v0.30.0 // indirect
-	google.golang.org/protobuf v1.36.10 // indirect
-)
