+++ remotewrite/sender/main_test.go
@@ -11,59 +11,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package sender
 
 import (
-	"fmt"
-	"os"
-	"strings"
 	"testing"
 	"time"
-
-	"github.com/prometheus/compliance/remotewrite/sender/targets"
 )
-
-const senderEnvVar = "PROMETHEUS_COMPLIANCE_RW_SENDERS"
-
-var (
-	// registeredTargets holds pre-defined targets to choose from.
-	//
-	// Custom targets could be considered for adding here, however the process target likely offers enough flexibility.
-	registeredTargets = map[string]targets.Target{
-		"prometheus": targets.RunPrometheus, // Default if no senderEnvVar is specified.
-		"process":    targets.RunProcess,
-	}
-	// targetsToTest is a global variable controlling senders to test.
-	// It is adjusted in TestMain via senderEnvVar variable.
-	targetsToTest = map[string]targets.Target{
-		"prometheus": registeredTargets["prometheus"],
-	}
-)
-
-// TestMain sets up the test environment by filtering registeredTargets (senders to tests) using
-// PROMETHEUS_COMPLIANCE_RW_SENDERS envvar.
-func TestMain(m *testing.M) {
-	senderNames := os.Getenv(senderEnvVar)
-	if senderNames != "" {
-		targetsToTest = make(map[string]targets.Target)
-		nameList := strings.Split(senderNames, ",")
-		for _, name := range nameList {
-			name = strings.TrimSpace(name)
-			if target, ok := registeredTargets[name]; ok {
-				targetsToTest[name] = target
-			} else {
-				fmt.Println("FAIL: Target from PROMETHEUS_COMPLIANCE_RW_SENDERS", name, "not registered")
-				os.Exit(1)
-			}
-		}
-		if len(targetsToTest) == 0 {
-			fmt.Println("FAIL: No targets found matching PROMETHEUS_COMPLIANCE_RW_SENDERS=", senderNames)
-			os.Exit(1)
-		}
-	}
-
-	os.Exit(m.Run())
-}
 
 // SenderTestScenario defines a test scenario to run against senders.
 // DEPRECATED: To kill, use sendertest.
@@ -77,7 +30,7 @@ type SenderTestScenario struct {
 
 // runSenderTest runs a test scenario using an automatic target.
 // DEPRECATED: To kill, use sendertest.
-func runSenderTest(t *testing.T, targetName string, target targets.Target, scenario SenderTestScenario) {
+func runSenderTest(t *testing.T, targetName string, target Sender, scenario SenderTestScenario) {
 	t.Helper()
 	t.Fatal("TODO: Remove")
 }
@@ -85,7 +38,7 @@ func runSenderTest(t *testing.T, targetName string, target targets.Target, scena
 // runAutoTargetWithCustomReceiver runs an auto-target with a custom receiver (for special test cases).
 // Use this when you need custom receiver behavior (e.g., TimestampTrackingReceiver, FallbackTrackingReceiver).
 // DEPRECATED: To kill, use sendertest.
-func runAutoTargetWithCustomReceiver(t *testing.T, targetName string, target targets.Target, receiverURL string, scrapeTarget *MockScrapeTarget, waitTime time.Duration) error {
+func runAutoTargetWithCustomReceiver(t *testing.T, targetName string, target Sender, receiverURL string, scrapeTarget *MockScrapeTarget, waitTime time.Duration) error {
 	t.Helper()
 
 	t.Fatal("TODO: Remove")
@@ -93,6 +46,6 @@ func runAutoTargetWithCustomReceiver(t *testing.T, targetName string, target tar
 }
 
 // forEachSender runs the provided test function for each configured sender.
-func forEachSender(t *testing.T, f func(*testing.T, string, targets.Target)) {
+func forEachSender(t *testing.T, f func(*testing.T, string, Sender)) {
 	t.Fatal("TODO: Remove")
 }
