+++ remotewrite/sender/protocol_test.go
@@ -11,14 +11,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package sender
 
 import (
 	"fmt"
 	"strings"
 	"testing"
-
-	"github.com/prometheus/compliance/remotewrite/sender/targets"
 )
 
 // TestProtocolCompliance validates HTTP protocol requirements for Remote Write 2.0 senders.
@@ -138,7 +136,7 @@ func TestHTTPMethod_Old(t *testing.T) {
 	t.Attr("rfcLevel", "MUST")
 	t.Attr("description", "Sender MUST use POST method for remote write requests")
 
-	forEachSender(t, func(t *testing.T, targetName string, target targets.Target) {
+	forEachSender(t, func(t *testing.T, targetName string, target Sender) {
 		runSenderTest(t, targetName, target, SenderTestScenario{
 			ScrapeData: "test_metric 42\n",
 			Validator: func(t *testing.T, req *CapturedRequest) {
